Refactor toggleCheckbox
